remove HiveDDLCommandSuite.scala
